backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv
apscheduler
pydantic-settings
